Move UniqueEntityList one level higher
UniqueEntityList is no longer UniqueOwnerList, and therefore should not
be parked under the Owner package.
