+++ src/main/java/dog/pawbook/model/AddressBook.java
@@ -5,7 +5,7 @@ import static java.util.Objects.requireNonNull;
 import java.util.List;
 
 import dog.pawbook.model.managedentity.Entity;
-import dog.pawbook.model.managedentity.owner.UniqueEntityList;
+import dog.pawbook.model.managedentity.UniqueEntityList;
 import javafx.collections.ObservableList;
 import javafx.util.Pair;
